refactor, added compactNum in calorie count

+++ app/src/main/java/com/example/simplecalorietracker/model/entity/CourseCard.kt
@@ -13,5 +13,5 @@ data class FoodEntry(
     val foodName: String,
 
     @SerializedName("calorieCount")
-    val calorieCount: Int,
+    val calorieCount: Long,
 )
